refactor: move to make it clear that this is a nounspace managed farcaster signer
